// types/Tender.ts
export type Tender = {
  _id: string;
  tenderNo: string;
  description: string;
  tenderDate: string;
  tenderType: string;
  status: string;
  position: string;
  tenderingDepartment: string;
  client: string;
  tenderValue: number;
  emdAmount: number;
  bgAmount: number;
  emdRefundDate?: string;
  emdRefundStatus?: string;
  bgRefundStatus?: string;
  owner?: string;
  remarks?: string;
  vertical?: string;
  subVertical?: string;
  country?: string;
  state?: string;
  tenderOpeningDate?: string;
  preBidMeetingDate?: string;
  tenderSubmissionLastDate?: string;
  documentFee?: number;
  transactionFee?: number;
  corpusFund?: number;
  isMAFRequired?: boolean;
  clientId?: string;
  tenderManager?: string;
  tenderManagerEmail?: string;
  tenderManagerPhone?: string;
  tenderOwner?: string;
  scm?: string;
};